rss_aggregate/requirements.txt
feedparser>=6.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
lxml>=4.6.0
aiohttp>=3.7.0
PyYAML>=5.4.0